backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
Pillow==10.1.0
pdf2image==1.16.3
reportlab==4.0.7
hypothesis==6.92.1
pytest==7.4.3
pytest-cov==4.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
